+++ drivers/timer/Kconfig
@@ -189,7 +189,6 @@ config RV32M1_LPTMR_TIMER
 	bool "RV32M1 LPTMR system timer driver"
 	default y
 	depends on SOC_OPENISA_RV32M1_RISCV32
-	depends on !TICKLESS_IDLE
 	depends on RV32M1_INTMUX
 	help
 	  This module implements a kernel device driver for using the LPTMR
@@ -199,7 +198,6 @@ config RV32M1_LPTMR_TIMER
 config LITEX_TIMER
 	bool "LiteX Timer"
 	default y
-	depends on !TICKLESS_IDLE
 	depends on SOC_RISCV32_LITEX_VEXRISCV
 	help
 	  This module implements a kernel device driver for LiteX Timer.
