+++ tests/bluetooth/bsim_bt/edtt_ble_test_app/hci_test_app/prj.conf
@@ -23,5 +23,3 @@ CONFIG_BT_CTLR_TX_BUFFER_SIZE=60
 CONFIG_BT_CTLR_DATA_LENGTH_MAX=60
 
 CONFIG_PM=y
-CONFIG_TICKLESS_IDLE=y
-CONFIG_TICKLESS_KERNEL=y
