--- tests/kernel/tickless/tickless/prj.conf
@@ -1,5 +1,0 @@
-CONFIG_PM=y
-CONFIG_ZTEST=y
-
-# The code is written to assume a slow tick rate
-CONFIG_SYS_CLOCK_TICKS_PER_SEC=100
