--- tests/kernel/tickless/tickless/testcase.yaml
@@ -1,6 +1,0 @@
-tests:
-  kernel.tickless:
-    arch_exclude: nios2 riscv32
-    filter: CONFIG_X86 or (CONFIG_ARM and (CONFIG_SOC_MK64F12
-      or CONFIG_SOC_SERIES_SAM3X)) or CONFIG_ARCH_POSIX
-    tags: tickless kernel
